pyproject.toml
[project]
name = "lpr-detect"
version = "0.1.0"
description = "yolo based lpr detection"
authors = [
    {name = "Anton",email = "antoxaz18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"


[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
numpy = "^2"
torch = { version = "^2.6", source = "torch118" }
torchvision = { version = "*", source = "torch118" }
torchaudio = { version = "*", source = "torch118" }
pyyaml = "^6.0.2"
tqdm = "^4.67.1"
pybind11 = "^2.13.6"
wheel = "^0.45.1"
setuptools = "^80.9.0"
onnx = "^1.18.0"
onnxscript = "^0.3.0"
onnxoptimizer = "^0.3.13"
onnx-simplifier = "^0.4.36"
opencv-python = "^4.11.0.86"
onnxruntime = "^1.22.0"



[[tool.poetry.source]]
name = "torch118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
